// src/app/models/Client.ts
export interface Cliente {

  id?: number;

  nombreCliente: string;
  apellidoCliente: string;
  nitCliente: string;
  direccionCliente: string;
  correoCliente: string;
  telefonoCliente: string;
  telefonoSecundarioCliente: string;
  estadoCliente: string;
  ciudadCliente: string;
  empresaCliente: string;
  caracteristicaCliente: string;

  generoUsuario: number;

}